extract progress recording into cli package
not used yet, but plan to use it for displaying progress in the web UI
