+++ pkg/runtimes/buildkit.go
@@ -1,7 +1,6 @@
 package runtimes
 
 import (
-	"bytes"
 	"context"
 	"crypto/sha256"
 	"embed"
@@ -12,8 +11,6 @@ import (
 	"os"
 	"path"
 	"path/filepath"
-	"regexp"
-	"sort"
 	"strings"
 	"sync"
 	"text/tabwriter"
@@ -33,12 +30,10 @@ import (
 	"github.com/moby/buildkit/session/secrets/secretsprovider"
 	"github.com/moby/buildkit/util/entitlements"
 	"github.com/morikuni/aec"
-	"github.com/opencontainers/go-digest"
 	ocispecs "github.com/opencontainers/image-spec/specs-go/v1"
-	perrors "github.com/pkg/errors"
-	"github.com/segmentio/textio"
 	"github.com/tonistiigi/units"
 	"github.com/vito/bass/pkg/bass"
+	"github.com/vito/bass/pkg/cli"
 	"github.com/vito/bass/pkg/ioctx"
 	"github.com/vito/progrock"
 	"github.com/vito/progrock/graph"
@@ -808,10 +803,9 @@ func (nopCloser) Close() error { return nil }
 
 func forwardStatus(rec *progrock.Recorder) *statusProxy {
 	return &statusProxy{
-		rec: rec,
-		wg:  new(sync.WaitGroup),
-		vs:  map[digest.Digest]*vertex{},
-		vsL: new(sync.Mutex),
+		rec:  rec,
+		wg:   new(sync.WaitGroup),
+		prog: cli.NewProgress(),
 	}
 }
 
@@ -819,10 +813,9 @@ func forwardStatus(rec *progrock.Recorder) *statusProxy {
 // status messages, and also records the progress so that we can emit it in a
 // friendlier error message
 type statusProxy struct {
-	rec *progrock.Recorder
-	wg  *sync.WaitGroup
-	vs  map[digest.Digest]*vertex
-	vsL *sync.Mutex
+	rec  *progrock.Recorder
+	wg   *sync.WaitGroup
+	prog *cli.Progress
 }
 
 func (proxy *statusProxy) proxy(rec *progrock.Recorder, statuses chan *kitdclient.SolveStatus) {
@@ -863,7 +856,7 @@ func (proxy *statusProxy) proxy(rec *progrock.Recorder, statuses chan *kitdclien
 			Logs:     ls,
 		}
 
-		proxy.Record(gstatus)
+		proxy.prog.WriteStatus(gstatus)
 		rec.Record(gstatus)
 	}
 }
@@ -880,182 +873,6 @@ func (proxy *statusProxy) Writer() chan *kitdclient.SolveStatus {
 	return statuses
 }
 
-type vertex struct {
-	*graph.Vertex
-
-	Log *bytes.Buffer
-}
-
-func (proxy *statusProxy) Record(status *graph.SolveStatus) {
-	proxy.vsL.Lock()
-	defer proxy.vsL.Unlock()
-
-	for _, v := range status.Vertexes {
-		ver, found := proxy.vs[v.Digest]
-		if !found {
-			ver = &vertex{Log: new(bytes.Buffer)}
-			proxy.vs[v.Digest] = ver
-		}
-
-		ver.Vertex = v
-	}
-
-	for _, l := range status.Logs {
-		ver, found := proxy.vs[l.Vertex]
-		if !found {
-			continue
-		}
-
-		_, _ = ver.Log.Write(l.Data)
-	}
-}
-
 func (proxy *statusProxy) NiceError(msg string, err error) bass.NiceError {
-	return &progrockErr{
-		msg: msg,
-		err: err,
-		vs:  proxy.vs,
-	}
-}
-
-type progrockErr struct {
-	msg string
-	err error
-	vs  map[digest.Digest]*vertex
-}
-
-func (err progrockErr) Error() string {
-	rootErr := perrors.Cause(err.err)
-	return fmt.Sprintf("%s: %s", err.msg, stripUselessPart(rootErr.Error()))
-}
-
-func (err progrockErr) NiceError(w io.Writer) error {
-	fmt.Fprintf(w, aec.RedF.Apply("%s")+"\n", err.Error())
-	fmt.Fprintln(w)
-
-	vtxPrinter{
-		vs:      err.vs,
-		printed: map[digest.Digest]struct{}{},
-	}.printAll(textio.NewPrefixWriter(w, "  "))
-
-	fmt.Fprintln(w)
-	fmt.Fprintln(w, aec.YellowF.Apply("for more information, refer to the full output above"))
-
-	return nil
-}
-
-func duration(dt time.Duration) string {
-	prec := 1
-	sec := dt.Seconds()
-	if sec < 10 {
-		prec = 2
-	} else if sec < 100 {
-		prec = 1
-	}
-
-	return fmt.Sprintf(aec.Faint.Apply("[%.[2]*[1]fs]"), sec, prec)
-}
-
-type vtxPrinter struct {
-	vs      map[digest.Digest]*vertex
-	printed map[digest.Digest]struct{}
-}
-
-func (printer vtxPrinter) printAll(w io.Writer) {
-	byStartTime := make([]*vertex, 0, len(printer.vs))
-	for _, vtx := range printer.vs {
-		byStartTime = append(byStartTime, vtx)
-	}
-	sort.Slice(byStartTime, func(i, j int) bool {
-		if byStartTime[i].Started != nil && byStartTime[j].Started == nil {
-			return true
-		}
-
-		if byStartTime[i].Started == nil && byStartTime[j].Started != nil {
-			return false
-		}
-
-		if byStartTime[i].Started == nil && byStartTime[j].Started == nil {
-			return byStartTime[i].Name < byStartTime[j].Name
-		}
-
-		return byStartTime[i].Started.Before(*byStartTime[j].Started)
-	})
-
-	for _, vtx := range byStartTime {
-		printer.print(w, vtx)
-	}
-}
-
-const maxLines = 24
-
-func (printer vtxPrinter) print(w io.Writer, vtx *vertex) error {
-	if _, printed := printer.printed[vtx.Digest]; printed {
-		return nil
-	}
-
-	printer.printed[vtx.Digest] = struct{}{}
-
-	if strings.Contains(vtx.Name, "[hide]") {
-		return nil
-	}
-
-	for _, input := range vtx.Inputs {
-		iv, found := printer.vs[input]
-		if !found {
-			continue
-		}
-
-		printer.print(w, iv)
-	}
-
-	if vtx.Cached {
-		fmt.Fprintf(w, aec.BlueF.Apply("=> %s"), vtx.Name)
-	} else if vtx.Error != "" {
-		if strings.HasSuffix(vtx.Error, context.Canceled.Error()) {
-			fmt.Fprintf(w, aec.YellowF.Apply("=> %s [canceled]")+" %s", vtx.Name, duration(vtx.Completed.Sub(*vtx.Started)))
-		} else {
-			fmt.Fprintf(w, aec.RedF.Apply("=> %s")+" %s", vtx.Name, duration(vtx.Completed.Sub(*vtx.Started)))
-		}
-	} else if vtx.Started == nil {
-		fmt.Fprintf(w, aec.Faint.Apply("=> %s"), vtx.Name)
-	} else if vtx.Completed != nil {
-		fmt.Fprintf(w, aec.GreenF.Apply("=> %s")+" %s", vtx.Name, duration(vtx.Completed.Sub(*vtx.Started)))
-	} else {
-		fmt.Fprintf(w, aec.YellowF.Apply("=> %s"), vtx.Name)
-	}
-
-	fmt.Fprintln(w)
-
-	if vtx.Error != "" {
-		if vtx.Log.Len() > 0 {
-			trimTrail := bytes.TrimRight(vtx.Log.Bytes(), "\n")
-			lines := bytes.Split(trimTrail, []byte("\n"))
-			if len(lines) > maxLines {
-				extra := len(lines) - maxLines
-				lines = lines[len(lines)-maxLines-1:]
-				fmt.Fprintf(w, aec.Faint.Apply("... %d lines omitted ...")+"\n", extra)
-			}
-
-			for _, line := range lines {
-				fmt.Fprintln(w, string(line))
-			}
-		}
-
-		if !isCanceled(vtx.Error) {
-			fmt.Fprintf(w, aec.RedF.Apply("ERROR: %s")+"\n", stripUselessPart(vtx.Error))
-		}
-	}
-
-	return nil
-}
-
-var processErr = regexp.MustCompile(`process ".*" did not complete successfully: `)
-
-func stripUselessPart(msg string) string {
-	return processErr.ReplaceAllString(msg, "")
-}
-
-func isCanceled(msg string) bool {
-	return strings.HasSuffix(msg, context.Canceled.Error())
+	return proxy.prog.WrapError(msg, err)
 }
